% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{check_overlap}
\alias{check_overlap}
\title{check_overlap}
\usage{
check_overlap(mod)
}
\arguments{
\item{mod}{a fitted model with fit_torpor}
}
\value{
a numeric value
}
\description{
The function check_overlap() generates prior posterior overlap values for Tm,
Tt and Betae.
Note: Values larger than 0.3 should lead to the conclusion that conforming
torpor, regulated torpor or thermoregulation respectively could not be modeled
with the data provided (Fasel et al. (in prep)).
}
\examples{
#'data(test_data2)
test <- fit_torpor(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.8,
Model = NULL,
fitting_options = list(nc = 1))
check_overlap(mod = test)
}
