% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{.step_4_bis}
\alias{.step_4_bis}
\title{step_4_bis}
\usage{
.step_4_bis(out_assignation, fitting_options, threshold)
}
\arguments{
\item{out_assignation}{output of estimatate_assignation.}

\item{fitting_options}{fitting options}

\item{threshold}{Assignment confidence threshold}
}
\description{
This function flag the uncertain point based on a threshold
}
