% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_assignation}
\alias{estimate_assignation}
\title{estimate_assignation}
\usage{
estimate_assignation(
  Y,
  Ta,
  bmr = NULL,
  tlc = NULL,
  fitting_options = list(ni = 50000, nt = 10, nb = 20000, nc = 2)
)
}
\arguments{
\item{bmr}{bmr value if not estimated}

\item{tlc}{tlc value if not estimated}

\item{fitting_options}{option to be passed to rjags.}
}
\description{
This function estimate the lowest TLC and the bmr
}
\examples{
\dontrun{
t2_no_input <- estimate_assignation(Ta = test_data2$Ta, Y = test_data2$VO2)
}
}
