% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{estimate_parameters}
\usage{
estimate_parameters(
  Ta,
  Y,
  bmr = NULL,
  tlc = NULL,
  fitting_options = list(ni = 50000, nt = 10, nb = 30000, nc = 3)
)
}
\arguments{
\item{Ta}{A vector of ambient temperature.}

\item{Y}{A vector of methabolic measure.}

\item{bmr}{bmr value if not estimated}

\item{tlc}{tlc value if not estimated}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\description{
\code{\link[=estimate_parameters]{estimate_parameters()}} fits a binomial mixture model using Bayesian
inference.  The function considers the assumed relation between metabolic rate (MR) and ambient temperature (Ta)
(Speakman & Thomas 2003). In the hypothermic state (torpor) and above some
threshold Ta (Tmin), MR follows an exponential curve reflecting the Arrhenius
rate enhancing effect of temperature on chemical reactions, whereas below
Tmin, it increases linearly with decreasing Ta to maintain a minimal body temperature (Tb) in
torpor. In the euthermic state, MR solely increases linearly with decreasing Ta.
The function uses Rjags in the background and enables users to specify
some - but not all - sampling parameters. The structure of the model can also
be changed. Users who want more flexibility are encouraged to use Rjags
directly.
More information about the model can be found in the \code{vignette("model_description", package = "torpor")}.
}
\details{
This model should be applied with sufficient sample size and if
evidence suggest that individuals under study will conform to the previously
described pattern while in torpor.
}
\examples{
\dontrun{
test_mod <- estimate_parameters(Ta = test_data3$Ta, Y = test_data3$VO2)
}
}
