% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_tlc_mtnz}
\alias{estimate_tlc_mtnz}
\title{Estimate Tlc}
\usage{
estimate_tlc_mtnz(
  M,
  Ta,
  MTNZ,
  fitting_options = list(ni = 50000, nt = 10, nb = 20000, nc = 2, parallel = TRUE)
)
}
\arguments{
\item{M}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{MTNZ}{A value of MTNZ.}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\description{
This function estimate the lowest Tlc and the MTNZ
}
\examples{
t <- estimate_tlc_mtnz(M = test_data2$VO2ms, Ta = test_data2$Ta, MTNZ = NULL)
}
