% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{eut_predict_fun}
\alias{eut_predict_fun}
\title{eut_predict_fun}
\usage{
eut_predict_fun(x, inte, betat, Ym)
}
\arguments{
\item{x}{a temperature}

\item{inte}{intercept 1}

\item{betat}{slope 1}

\item{Ym}{mean of Y to back transform}
}
\value{
a metabolic value
}
\description{
function to fit the model in euthermy // used internally only
}
