% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{find_low_tlc_bmr}
\alias{find_low_tlc_bmr}
\title{find_low_tlc_bmr}
\usage{
find_low_tlc_bmr(Y, Ta)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}
}
\value{
A named vector with bmr and low_tlc.
}
\description{
This function estimate the lowest TLC and the bmr
}
\examples{
t_1 <- find_low_tlc_bmr(Y = test_data2$VO2ms, Ta = test_data2$Ta)
}
