% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{find_low_Tlc_Mtnz}
\alias{find_low_Tlc_Mtnz}
\title{Find the low Tlc.}
\usage{
find_low_Tlc_Mtnz(M, Ta)
}
\arguments{
\item{M}{A vector of metabolic rate measure.}

\item{Ta}{A vector of ambient temperature.}
}
\value{
A named vector with Mtmz and low_Tlc.
}
\description{
This function estimate the lowest Tlc and the Mtnz
}
\examples{
t_1 <- find_low_Tlc_Mtnz(M = test_data$VO2, Ta = test_data$Ta)
}
