% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fit_and_plot}
\alias{fit_and_plot}
\title{plot raw data and model output}
\usage{
fit_and_plot(data, model_out = NULL, ...)
}
\arguments{
\item{data}{a data frame with 2 column MR and Ta}

\item{model_out}{a fitted model from fit torpor}

\item{...}{arguments to fit the model in \code{\link{fit_torpor}} if not provided}
}
\value{
a plot
}
\description{
plot the raw values and the fitted model
}
