% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{fit_and_plot2}
\alias{fit_and_plot2}
\title{plot raw data and model output}
\usage{
fit_and_plot2(mod = NULL, plot_type = "ggplot", MR, Ta, ...)
}
\arguments{
\item{mod}{a fitted model from fit torpor}

\item{plot_type}{character string specifying the type of plot desired. either "base" or "ggplot"}

\item{MR}{a vector with MR}

\item{Ta, }{a vector with Ta (same length as MR)}

\item{...}{arguments to fit the model in \code{\link[=fit_torpor]{fit_torpor()}} if no model output are provided}
}
\value{
a base-R plot or a ggplot object
}
\description{
The function provides a plot of the MR values over the respective Ta.
Measures are presented in different colors depending of the metabolic
stage and regression slopes (continuous lines) as well as 95% CI (segmented lines) are presented.
For more flexibility the users are advised to use \code{\link[=fit_torpor]{fit_torpor()}} and \code{\link[=get_prediction]{get_prediction()}} directly.
}
\examples{
\dontrun{
data(test_data)
fit_and_plot(MR = test_data[,2], Ta = test_data[,1], BMR = 29, TLC = 28.8,
fitting_options = list(nc =1), plot_type = "ggplot")
}
}
