% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fit_model}
\alias{fit_model}
\alias{fit_torpor}
\title{Fit models}
\usage{
fit_torpor(data, BMR, TLC, Model = NULL, fitting_options = list(ni = 5e+05,
  nt = 10, nb = 3e+05, nc = 3))
}
\arguments{
\item{data}{a data frame with 2 col. 1: Temperature, & 2: Metabolic value}

\item{BMR}{value for the focal specie}

\item{TLC}{value for the focal specie}

\item{Model}{path to model file}

\item{fitting_options}{a list with specification for jags parameters}
}
\value{
a fitted jags object
}
\description{
wrapper around the Jags function -
matic rounding
}
