% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{funnorm}
\alias{funnorm}
\title{funnorm}
\usage{
funnorm(x, int1, beta1)
}
\arguments{
\item{x}{a temperature}

\item{int1}{intercept 1}

\item{beta1}{slope 1}
}
\value{
a metabolic value
}
\description{
fit the model for normotermic bats
}
