% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{funtorp2}
\alias{funtorp2}
\title{funtorp2}
\usage{
funtorp2(x, Tt, intr, intc, betat, betac, Ym)
}
\arguments{
\item{x}{a temperature}

\item{Tt}{turning point T}

\item{intr}{intercept 1}

\item{intc}{intercept 2}

\item{betat}{slope 1}

\item{betac}{slope 2}

\item{Ym}{mean of Y to back-transform}
}
\value{
a metabolic value
}
\description{
fit the mod for torpor bats
}
