% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction2.R
\name{get_classification}
\alias{get_classification}
\title{Get classification}
\usage{
get_classification(mod)
}
\arguments{
\item{mod}{a fitted model from fit_torpor}
}
\value{
a data frame with classification and predicted MR
}
\description{
The function returns a functions with the classification (torpor or euthermy)
for the observed data. It also provides the mean predicted values of MR at the
given Ta.
}
\examples{
\dontrun{
data(test_data2)
test_mod <- fit_torpor(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1))
get_classification(mod = test_mod)
}
}
