% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction2.R
\name{get_prediction}
\alias{get_prediction}
\title{Get Prediction}
\usage{
get_prediction(mod, Ta)
}
\arguments{
\item{mod}{a fitted model from fit_torpor}

\item{Ta}{a vector of temperatur for which the prediction should be made}
}
\value{
a data frame with predicted values
}
\description{
The function provides the MR estimates and 95% credible interval boundaries
at a defined Ta in normothermic and/or torpid stage.
}
\examples{
\dontrun{
data(test_data2)
test_mod <- fit_torpor(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1))
get_prediction(mod = test_mod, Ta = 1:20)
}
}
