% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction2.R
\name{get_prediction2}
\alias{get_prediction2}
\title{Get Prediction2}
\usage{
get_prediction2(mod, Ta)
}
\arguments{
\item{mod}{a fitted model from fit_torpor}

\item{Ta}{a vector of temperatur for which the prediction should be made}
}
\value{
a data frame with predicted values
}
\description{
The function provides the MR estimates and 95% credible interval boundaries
at a defined Ta in normothermic and/or torpid stage.
}
\examples{
\dontrun{
data(test_data2)
test_mod <- fit_torpor2(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 98,
TLC = 28.88,
mod = NULL,
fitting_options = list(nc = 1))
get_prediction2(mod = test_mod, Ta = 1:20)
}
}
