% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q10.R
\name{get_Q10}
\alias{get_Q10}
\title{get_Q10}
\usage{
get_Q10(mod)
}
\arguments{
\item{mod}{a fitted model from fit_torpor}
}
\value{
the Q10 value
}
\description{
The function provides the Q10 temperature coefficient,
which represents the metabolic rates ratio measured at 10 °C Ta apart.
}
