% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{get_summary}
\usage{
get_summary(mod)
}
\arguments{
\item{mod}{a fitted model with fit_torpor2}
}
\value{
a list
}
\description{
This function return a comprehensive summary of the model fit. It returns the
estimated parameters of the model. The check of overlap, Q10,
}
\examples{
data(test_data2)
test2 <- fit_torpor2(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.8,
Model = NULL,
fitting_options = list(nc = 1))
get_summary(mod = test2)
}
