% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hinit_hetero}
\alias{hinit_hetero}
\alias{inits_hetero}
\title{init hetero used internally only}
\usage{
inits_hetero(x)
}
\arguments{
\item{x}{TLC value for the focal species}
}
\value{
a list with initial values
}
\description{
matic rounding
}
