% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_torpor}
\alias{plot_torpor}
\title{plot raw data and model output}
\usage{
plot_torpor(mod = NULL, plot_type = "ggplot",
  col_torp = "cornflowerblue", col_eut = "coral3", ylab = "M",
  xlab = "Ta", pdf = FALSE, ...)
}
\arguments{
\item{mod}{a fitted model from \code{\link[=fit_torpor]{fit_torpor()}}}

\item{plot_type}{A character string specifying the type of plot desired. Either "base" or "ggplot"}

\item{col_torp}{color for torpor model fit and points}

\item{col_eut}{color for euthermy model fit and points}

\item{ylab}{y label}

\item{xlab}{x label}

\item{pdf}{logical if a .pdf copy of the plot should be saved}

\item{...}{arguments to fit the model in \code{\link[=fit_torpor]{fit_torpor()}} if no model is provided}
}
\value{
a base-R plot or a ggplot object
}
\description{
The function provides a plot of the MR values over the respective Ta.
Measures are presented in different colors depending of the metabolic
stage and regression slopes (continuous lines) as well as 95% CI (segmented lines) are presented.
For more flexibility the users are advied to use \code{\link[=fit_torpor]{fit_torpor()}} and \code{\link[=get_prediction]{get_prediction()}} directly.
}
\examples{
data(test_data3)
plot_torpor(MR = test_data3[,2], Ta = test_data3[,1], BMR = 1.055, TLC = 29,
fitting_options = list(nc =1), plot_type = "ggplot")
}
