% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{step_2}
\alias{step_2}
\alias{estimate_tlc_bmr}
\title{step_2}
\usage{
estimate_tlc_bmr(
  Y,
  Ta,
  fitting_options = list(ni = 50000, nt = 10, nb = 20000, nc = 2)
)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{out_1}{output of step_1.}
}
\description{
This function estimate the lowest TLC and the bmr
}
\examples{
t <- step_2(test_data2$VO2ms, test_data2$Ta)
t2 <- step_2(Y = test_data2$VO2ms, Ta = test_data2$Ta, out_1 = t_1)
}
