% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test_data}
\format{
A data frame with 28 obs. taken from Genoud EA 1997
}
\usage{
data(test_data)
}
\description{
Callithrix_pygmaea
BMR: 98
TLC: 28.88
}
\keyword{datasets}
