% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data2}
\alias{test_data2}
\title{test_data2}
\format{A data frame with 103 obs. taken from Geiser 1987}
\usage{
data(test_data2)
}
\description{
Cercartetus_lepidus
BMR: 1.49
TLC: 28.8
}
\keyword{datasets}
