% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data3}
\alias{test_data3}
\title{test_data3}
\format{
A data frame with 75 obs. Geiser & Baudinette 1987
}
\usage{
data(test_data3)
}
\description{
Dasyuroides byrnei
Mtnz = 0.7, Tlc = 28.3
}
\keyword{datasets}
