% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_assign}
\alias{tor_assign}
\title{Assign metabolic state}
\usage{
tor_assign(tor_obj)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}
}
\value{
a data.frame
}
\description{
\code{\link[=tor_assign]{tor_assign()}} The function assign the individual points according to their
estimated state.
}
\seealso{
Other predict: 
\code{\link{eut_predict_fun}()},
\code{\link{tor_predict_fun}()},
\code{\link{tor_predict}()}
}
\concept{predict}
