% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_classify}
\alias{tor_classify}
\title{tor_classify}
\usage{
tor_classify(mod)
}
\arguments{
\item{mod}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}
}
\value{
a data.frame
}
\description{
\code{\link[=tor_classify]{tor_classify()}} returns the raw data with the related
predicted state values (between 1 and 2), which leads to the state
classification (torpor or euthermy). Additionally, it also provides the
predicted metabolic rate (MR) at the given ambient temperature (Ta).
}
\examples{
data(test_data2)
test_mod <- tor_fit(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1, ni = 5000, nb = 3000))
tor_classify(mod = test_mod)
}
\seealso{
Other predict: 
\code{\link{eut_predict_fun}()},
\code{\link{tor_predict_fun}()},
\code{\link{tor_predict}()}
}
\concept{predict}
