% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{tor_overlap}
\alias{tor_overlap}
\title{tor_overlap}
\usage{
tor_overlap(mod)
}
\arguments{
\item{mod}{a fitted model with fit_torpor}
}
\value{
a list of 3. With overlapping values for tlc, Tt and Betat
}
\description{
The function tor_overlap() generates prior/posterior overlap values for Tm,
Tt and Betat.
Note: Values larger than 0.3 should lead to the conclusion that conforming
torpor, regulated torpor or thermoregulation respectively could not be modeled
with the data provided (Fasel et al. (in prep)). Can be used independently but is also used internally in
tor_summarise()
}
\examples{
data(test_data2)
test <- tor_fit(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.8,
model = NULL,
fitting_options = list(nc = 1, ni = 5000, nb = 3000))
tor_overlap(mod = test)
}
\seealso{
Other summary: \code{\link{get_overlap}},
  \code{\link{get_parameters}}, \code{\link{tor_summarise}}
}
\concept{summary}
