% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_plot.R
\name{tor_plot}
\alias{tor_plot}
\title{Plot raw data and predicted values of a fitted model}
\usage{
tor_plot(
  tor_obj = NULL,
  plot_type = "base",
  col_torp = "cornflowerblue",
  col_euth = "coral3",
  col_Mtnz = "black",
  ylab = "M",
  xlab = "Ta",
  xlim = NULL,
  ylim = NULL,
  legend = TRUE,
  pdf = FALSE
)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}

\item{plot_type}{A character string specifying the type of plot desired. Either "base" or "ggplot". Note that ggplots are still in development stage.}

\item{col_torp}{color for torpor model prediction and points}

\item{col_euth}{color for Euthermia model prediction and points}

\item{col_Mtnz}{color of Mtnz}

\item{ylab}{y label}

\item{xlab}{x label}

\item{xlim}{limit of the x-axis}

\item{ylim}{limit of the y-axis}

\item{legend}{A logical specifying if a legend should be added to the plot. Only work for base plot at the moment.}

\item{pdf}{logical if a .pdf copy of the plot should be saved}
}
\value{
a graphics plot or a ggplot object
}
\description{
\code{\link[=tor_plot]{tor_plot()}} provides a plot of the metabolic rate (M) values over the respective
ambient temperature (Ta).
}
\details{
The function plots the predicted value as well as the raw data.
Measures are presented in different colors depending on the metabolic state.
Predicted values as well as 95\% credible interval (segmented lines) are also presented.
This function enables the user to replicate the analysis done in Fasel et al. (in prep).
For more flexibility the users can use \code{\link[=tor_fit]{tor_fit()}} and \code{\link[=tor_predict]{tor_predict()}} directly.
}
