% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{tor_summarise}
\alias{tor_summarise}
\title{Model summary}
\usage{
tor_summarise(tor_obj)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=estimate_parameters]{estimate_parameters()}}}
}
\value{
a list of two data.frame. The first element returns the parameter estimates, the second reports the overlap values.
}
\description{
\code{\link[=tor_summarise]{tor_summarise()}} provides a comprehensive summary of the output
returned by \code{\link[=estimate_parameters]{estimate_parameters()}}. Values of mean, 95CI bounds, median and
Brooks–Gelman–Rubin criterion (i.e. chain convergence check) of parameters
posterior distributions are provided. Additionally, prior posterior overlap
values for parameters tlc, Tt and Betat are generated.
}
\examples{
data(test_data)
test2 <- estimate_parameters(Y = test_data[,2],Ta = test_data[, 1], fitting_options = list(nc = 1, ni = 5000, nb = 3000))
tor_summarise(tor_obj)
}
\seealso{
Other summary: 
\code{\link{get_parameters}()},
\code{\link{tor_overlap}()}
}
\concept{summary}
