% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_torpor.R
\name{complete_args}
\alias{complete_args}
\title{complete_args}
\usage{
complete_args(fn)
}
\arguments{
\item{fn}{a function}
}
\description{
this function completes the arguments of a function. Should not
be used directly by the user. Is used internally for fit_torpor
From Alex Courtiol
}
