% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{complete_args}
\alias{complete_args}
\title{complete_args // internal}
\usage{
complete_args(fn)
}
\arguments{
\item{fn}{a function}
}
\description{
this function is used internally for tor_fit
From Alex Courtiol - informed. He does not want to be listed as an author -
}
\seealso{
Other fit: \code{\link{tor_fit}}
}
\concept{fit}
