% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{.step_3_bis}
\alias{.step_3_bis}
\title{step_3_bis}
\usage{
.step_3_bis(out_assignation, fitting_options)
}
\arguments{
\item{fitting_options}{fitting options}

\item{out_2}{output of step_2}

\item{out_3}{output of step_3}
}
\description{
This function estimate the lowest TLC and the bmr
}
