% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_assignation}
\alias{estimate_assignation}
\title{Estimate assignation}
\usage{
estimate_assignation(
  M,
  Ta,
  Mtnz = NULL,
  Tlc = NULL,
  fitting_options = list(ni = 80000, nt = 10, nb = 30000, nc = 3, parallel = TRUE)
)
}
\arguments{
\item{M}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{Mtnz}{Mtnz value if not estimated}

\item{Tlc}{Tlc value if not estimated}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\description{
This function estimate the assignation into torpor or euthermy. It is used
internally prior to estimate the parameters
}
\examples{
\dontrun{
t2_no_input <- estimate_assignation(Ta = test_data2$Ta, M = test_data2$VO2)
to_input <- estimate_assignation(Ta = test_data2$Ta, M = test_data2$VO2, Mtnz = 1.49, Tlc = 28.8)
}
}
