% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_assignation}
\alias{estimate_assignation}
\title{estimate_assignation}
\usage{
estimate_assignation(
  Y,
  Ta,
  MTNZ = NULL,
  tlc = NULL,
  fitting_options = list(ni = 50000, nt = 10, nb = 20000, nc = 2, parallel = TRUE)
)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{tlc}{tlc value if not estimated}

\item{fitting_options}{a list of fitting option to pass to jags.}

\item{mtnz}{mtnz value if not estimated}
}
\description{
This function estimate the assignation into torpor or euthermy.
}
\examples{
\dontrun{
t2_no_input <- estimate_assignation(Ta = test_data2$Ta, Y = test_data2$VO2)
}
}
