% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_tlc_mtnz}
\alias{estimate_tlc_mtnz}
\title{estimate_tlc_mtnz}
\usage{
estimate_tlc_mtnz(
  Y,
  Ta,
  fitting_options = list(ni = 50000, nt = 10, nb = 20000, nc = 2, parallel = TRUE)
)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\description{
This function estimate the lowest TLC and the MTNZ
}
\examples{
t <- estimate_tlc_mtnz(test_data2$VO2ms, test_data2$Ta)
}
