% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{estimate_Tlc_Mtnz}
\alias{estimate_Tlc_Mtnz}
\title{Estimate Tlc}
\usage{
estimate_Tlc_Mtnz(
  M,
  Ta,
  Mtnz,
  fitting_options = list(ni = 50000, nt = 10, nb = 30000, nc = 3, parallel = TRUE)
)
}
\arguments{
\item{M}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{Mtnz}{A value of Mtnz.}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\description{
This function estimate the lowest Tlc and the Mtnz
}
\examples{
t <- estimate_Tlc_Mtnz(M = test_data2$VO2ms, Ta = test_data2$Ta, Mtnz = NULL,
                      fitting_options = list(parallel = FALSE))
}
