% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{find_low_tlc_mtnz}
\alias{find_low_tlc_mtnz}
\title{Find the low Tlc.}
\usage{
find_low_tlc_mtnz(M, Ta)
}
\arguments{
\item{M}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}
}
\value{
A named vector with mtmz and low_tlc.
}
\description{
This function estimate the lowest Tlc and the mtnz
}
\examples{
t_1 <- find_low_tlc_mtnz(M = test_data$VO2, Ta = test_data$Ta)
}
