% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fit_and_plot}
\alias{fit_and_plot}
\title{plot raw data and model output}
\usage{
fit_and_plot(model_out = NULL, MR, Ta, ...)
}
\arguments{
\item{model_out}{a fitted model from fit torpor}

\item{MR}{a vector with MR}

\item{Ta, }{a vector with Ta (same length as MR)}

\item{...}{arguments to fit the model in \code{\link{fit_torpor}} if not provided}
}
\value{
a plot
}
\description{
plot the raw values and the fitted model
}
\examples{
data(test_data)
fit_and_plot(MR = test_data[,2], Ta = test_data[,1], BMR = 98, TLC = 28.88,
fitting_options = list(nc =1) )
}
