% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_torpor.R
\name{fit_torpor}
\alias{fit_torpor}
\title{Fit Torpor}
\usage{
fit_torpor(MR, Ta, BMR, TLC, Model = NULL, fitting_options = list(ni =
  5e+05, nt = 10, nb = 3e+05, nc = 3))
}
\arguments{
\item{MR}{a vector of Metabolic rate}

\item{Ta}{a vector of Temperature (same length as Ta)}

\item{BMR}{value for the focal specie}

\item{TLC}{value for the focal specie}

\item{Model}{path to model_file.txt}

\item{fitting_options}{a list with specification for jags parameters}
}
\value{
a fitted jags object
}
\description{
wrapper around the function
}
\examples{
data(test_data)
}
