% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_torpor2.R
\name{fit_torpor}
\alias{fit_torpor}
\title{Fit Torpor}
\usage{
fit_torpor(MR, Ta, BMR, TLC, Model = NULL, fitting_options = list(ni =
  5e+05, nt = 10, nb = 3e+05, nc = 3))
}
\arguments{
\item{MR}{a vector of Metabolic rates}

\item{Ta}{a vector of ambient Temperatures (same length as MR)}

\item{BMR}{BMR value for the focal specie}

\item{TLC}{TLC value for the focal specie}

\item{Model}{path to model_file.txt if a different model is used}

\item{fitting_options}{a list specifying sampling parameters. The follwing parameters can be speficied:
\itemize{
\item ni = number of itterations
\item nt = thin rate
\item nb = number of burns
\item nc = number of chains
}}
}
\value{
jagsUI object
}
\description{
The function \code{fit_torpor} fit binomial mixture model using Bayesian inference.
It uses Rjags in the background and enables users to specify some - but not all -
sampling parameters. The structure of the model can also be changed. User who want more
flexibility are encouraged to use Rjags directly.
The function considers the assumed relation between MR and Ta (Speakman & Thomas 2003).
In the hypothermic state (torpor) and above some threshold Ta (Tmin),
MR follows an exponential curve reflecting the Arrhenius rate enhancing effect
of temperature on chemical reactions, whereas below Tmin, it increases linearly
with decreasing Ta to maintain a minimal Tb in torpor.
In the euthermic state, MR solely increases linearly with decreasing Ta.
CAUTION: This model should be applied only if enough evidence is available suggesting
that the individuals under study will conform to the previously described pattern while in torpor.
}
\examples{
rm(list = ls())
data(test_data3)
test <- fit_torpor(MR = test_data3[,2],
Ta = test_data3[, 1],
BMR = 1.005,
TLC = 29,
fitting_options = list(nc = 1))
}
