% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{funnorm}
\alias{funnorm}
\title{funnorm}
\usage{
funnorm(x, inte, betat, Ym)
}
\arguments{
\item{x}{a temperature}

\item{inte}{intercept 1}

\item{betat}{slope 1}

\item{Ym}{mean of Y to back transform}
}
\value{
a metabolic value
}
\description{
fit the mod for normotermic bats // used internally only
}
