% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{funtorp}
\alias{funtorp}
\title{funtorp}
\usage{
funtorp(x, Tmin, int2, int3, beta1, beta2)
}
\arguments{
\item{x}{a temperature}

\item{Tmin}{turning point T}

\item{int2}{intercept 1}

\item{int3}{intercept 2}

\item{beta1}{slope 1}

\item{beta2}{slope 2}
}
\value{
a metabolic value
}
\description{
fit the model for torpor bats
}
