% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction2.R
\name{get_classification}
\alias{get_classification}
\title{Get classification}
\usage{
get_classification(mod)
}
\arguments{
\item{mod}{a fitted model from fit_torpor}
}
\value{
a data frame with classification and predicted MR
}
\description{
The function get_classification() returns the raw data with the related
predicted stage values (between 1 and 2), which leads to the stage
classification (torpor or euthermy). Additionally, it also provides the
predicted MR at the given Ta.
}
\examples{
\dontrun{
data(test_data2)
test_mod <- fit_torpor(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1))
get_classification(mod = test_mod)
}
}
