% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{get_overlap}
\alias{get_overlap}
\title{get_overlap}
\usage{
get_overlap(mod, params, priors)
}
\arguments{
\item{mod}{a fitted model with fit_torpor}

\item{params}{a parameter of the model}

\item{priors}{a posterior distribution (as. numerical vector)}
}
\value{
a numeric value
}
\description{
get_overlap
}
