% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{get_overlap}
\alias{get_overlap}
\title{get_overlap}
\usage{
get_overlap(mod, params, priors)
}
\arguments{
\item{mod}{a fitted model with fit_torpor}

\item{params}{a parameter of the model}

\item{priors}{a posterior distribution (as. numerical vector)}
}
\value{
a numeric value
}
\description{
This function compare the distribution of the posterior with the prior for a
given parametters. It is used as a measure to decide which parameters has
been correctly identified. The threshold is set at 0.30. Used internally in
the function overlap_model
}
