% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{get_overlap}
\alias{get_overlap}
\title{get_overlap // internal}
\usage{
get_overlap(mod, params, priors)
}
\arguments{
\item{mod}{a fitted model with \code{\link[=tor_fit]{tor_fit()}}}

\item{params}{a parameter of the model}

\item{priors}{a posterior distribution (as. numerical vector)}
}
\value{
a numeric value
}
\description{
Compare the distribution overlap between a prior and a posterior // used internally only
}
\seealso{
Other summary: \code{\link{get_parameters}},
  \code{\link{tor_overlap}}, \code{\link{tor_summarise}}
}
\concept{summary}
