% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{get_parameters}
\alias{get_parameters}
\title{Fetch parameters from model output}
\usage{
get_parameters(tor_obj)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}
}
\value{
a data.frame
}
\description{
\code{\link[=get_parameters]{get_parameters()}} Retrieves the parameters estimates from the model output
and creates a dataframe used in the \code{\link[=tor_summarise]{tor_summarise()}} function
}
\seealso{
Other summary: 
\code{\link{tor_ppo}()},
\code{\link{tor_summarise}()}
}
\concept{summary}
