% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters // internal}
\usage{
get_parameters(mod)
}
\arguments{
\item{mod}{a fitted model with \code{\link[=tor_fit]{tor_fit()}}}
}
\value{
a data.frame
}
\description{
Retrieves the parameters estimates from the model output
it creates a dataframe used in the \code{\link[=tor_summarise]{tor_summarise()}} function
}
\seealso{
Other summary: \code{\link{get_overlap}},
  \code{\link{tor_overlap}}, \code{\link{tor_summarise}}
}
\concept{summary}
