% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{get_prediction}
\alias{get_prediction}
\title{Get Predictions}
\usage{
get_prediction(model, Temp)
}
\arguments{
\item{model}{a fitted model from fit_torpor}

\item{Temp}{a vector of temperatur for which the prediction should be made}
}
\value{
a data frame with predicted values
}
\description{
extract prediction from the output of the model
}
