% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction2.R
\name{get_prediction}
\alias{get_prediction}
\title{get Prediction}
\usage{
get_prediction(mod, Ta)
}
\arguments{
\item{mod}{a fitted model from the function \code{\link[=fit_torpor]{fit_torpor()}}}

\item{Ta}{a vector of temperatur for which the prediction should be made}
}
\value{
a data frame with predicted values
}
\description{
The function \code{\link[=get_prediction]{get_prediction()}} provides the predicted MR and 95CI bounds at
a given Ta, in normothermic and/or torpid stage.
}
\examples{
\dontrun{
data(test_data2)
test_mod <- fit_torpor(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1))
get_prediction(mod = test_mod, Ta = 1:20)
}
}
