% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{get_summary}
\usage{
get_summary(mod)
}
\arguments{
\item{mod}{a fitted model with fit_torpor}
}
\value{
a list
}
\description{
The function get_summary() provides a comprehensive summary of the output
returned by fit_torpor(). Values of mean, 95CI bounds, median and
Brooks–Gelman–Rubin criterion (i.e. chain convergence check) of parameters
posterior distributions are provided. Additionally, prior posterior overlap
values for different parameters are generated. Finally, the mean MR value used
to standardize the MR variable is reported.
}
\examples{
data(test_data)
test2 <- fit_torpor(MR = test_data[,2],
Ta = test_data[, 1],
BMR = 29,
TLC = 28.8,
Model = NULL,
fitting_options = list(nc = 1))
get_summary(mod = test2)
}
