% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{step_1}
\alias{step_1}
\title{step_1}
\usage{
step_1(MR, Ta)
}
\arguments{
\item{Ta}{A vector of ambient temperature.}

\item{Y}{A vector of methabolic measure.}
}
\description{
This function estimate the lowest TLC and the bmr
}
\examples{
step_1(test_data$VO2, test_data$Ta)
}
