% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{step_1_and_2}
\alias{step_1_and_2}
\title{step_1_and_2}
\usage{
step_1_and_2(
  Y,
  Ta,
  fitting_options = list(ni = 5e+05, nt = 10, nb = 3e+05, nc = 3)
)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{fitting_options}{a list specifying sampling parameters. The follwing parameters can be speficied:
\itemize{
\item ni = number of itterations: default ni = 500000
\item nt = thin rate: default nt = 10
\item nb = number of burns: default nb = 300000
\item nc = number of chains: default nc = 3
}}
}
\description{
This function estimate the lowest TLC and the bmr
}
