% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{step_3}
\alias{step_3}
\title{step_1_and_2}
\usage{
step_3(
  Y,
  Ta,
  bmr = NULL,
  tlc = NULL,
  fitting_options = list(ni = 5e+05, nt = 10, nb = 3e+05, nc = 3)
)
}
\arguments{
\item{Y}{A vector of methabolic measure.}

\item{Ta}{A vector of ambient temperature.}

\item{bmr}{bmr value if not estimated}

\item{tlc}{tlc value if not estimated}
}
\description{
This function estimate the lowest TLC and the bmr
}
\examples{
\dontrun{
step_3(Ta = test_data$Ta, Y = test_data$VO2)
}
}
