% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data3}
\alias{test_data3}
\title{test_data3}
\format{A data frame with 186 obs. taken from Lovegrove EA 2001}
\usage{
data(test_data3)
}
\description{
Elephantulus_rozeti
BMR: 1.055
TLC: 29
}
\keyword{datasets}
