% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_classify}
\alias{tor_classify}
\title{tor_classify}
\usage{
tor_classify(tor_obj)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}
}
\value{
a data.frame
}
\description{
\code{\link[=tor_classify]{tor_classify()}} returns the raw data with the related
predicted state values (between 1 and 2), which leads to the state
classification (torpor or euthermy). Additionally, it also provides the
predicted metabolic rate (MR) at the given ambient temperature (Ta).
}
\seealso{
Other predict: 
\code{\link{eut_predict_fun}()},
\code{\link{tor_predict_fun}()},
\code{\link{tor_predict}()}
}
\concept{predict}
