% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_classify}
\alias{tor_classify}
\title{tor_classify}
\usage{
tor_classify(mod)
}
\arguments{
\item{mod}{a fitted model from tor_fit}
}
\value{
a data.frame with classification and predicted MR
}
\description{
The function tor_classify() returns the raw data with the related
predicted stage values (between 1 and 2), which leads to the stage
classification (torpor or euthermy). Additionally, it also provides the
predicted MR at the given Ta.
}
\examples{
data(test_data2)
test_mod <- tor_fit(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 1, ni = 5000, nb = 3000))
tor_classify(mod = test_mod)
}
\seealso{
Other predict: \code{\link{eut_predict_fun}},
  \code{\link{tor_predict_fun}}, \code{\link{tor_predict}}
}
\concept{predict}
