% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{tor_fit}
\alias{tor_fit}
\title{Mixture model aimed at assigning metabolic rate measurements (M) to torpor and euthermia}
\usage{
tor_fit(
  Ta,
  M,
  MTNZ = NULL,
  Tlc = NULL,
  fitting_options = list(ni = 50000, nt = 10, nb = 30000, nc = 3, parallel = TRUE)
)
}
\arguments{
\item{Ta}{A vector of ambient temperature.}

\item{M}{A vector of methabolic measure.}

\item{MTNZ}{mtnz value if not estimated}

\item{Tlc}{tlc value if not estimated}

\item{fitting_options}{a list of fitting option to pass to jags.}
}
\value{
A list of class tor_obj
}
\description{
The function \code{\link[=tor_fit]{tor_fit()}} considers the relation between metabolic rate (M) and
ambient temperature (Ta) assumed by the Scholander-Irving model and its later extensions.
}
\details{
Resting M measured within the thermoneutral zone (TNZ) is independent of Ta.
This rate is hereafter referred to as MTNZ, although it would correspond to
the basal metabolic rate (BMR) provided that the specific criteria for the
BMR are met (see Fasel et al. in prep.). Below the lower critical temperature
of TNZ (Tlc), M of euthermic animals increases linearly with decreasing Ta.
M of torpid animals increases linearly with decreasing Ta to maintain a minimal
body temperature below some threshold ambient temperature (Tt). This state is
usually referred to as "regulated torpor". Between Tt and Tlc, M of torpid
animals follows an exponential curve. In this Ta range, torpor is referred to
as "conforming torpor".
}
\examples{
\dontrun{
test_mod <- tor_fit(Ta = test_data3$Ta, M = test_data3$Y,
                   fitting_options = list(parallel = TRUE))
test_mod2 <- tor_fit(Ta = test_data2$Ta, M = test_data2$VO2ms, MTNZ = 1.8, Tlc = 29.2,
                   fitting_options = list(parallel = TRUE))
test_mod3 <- tor_fit(Ta = test_data2$Ta, M = test_data2$VO2ms, MTNZ = 1.8,
                   fitting_options = list(parallel = TRUE))
}
}
