% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_fit.R
\name{tor_fit}
\alias{tor_fit}
\title{tor_fit}
\usage{
tor_fit(MR, Ta, BMR, TLC, model = NULL, fitting_options = list(ni =
  5e+05, nt = 10, nb = 3e+05, nc = 3))
}
\arguments{
\item{MR}{a vector of Metabolic rates}

\item{Ta}{a vector of ambient Temperatures (same length as MR)}

\item{BMR}{BMR value for the focal specie}

\item{TLC}{TLC value for the focal specie}

\item{model}{path to model_file.txt if a different model is used}

\item{fitting_options}{a list specifying sampling parameters. The follwing parameters can be speficied:
\itemize{
\item ni = number of itterations: default ni = 500000
\item nt = thin rate: default nt = 10
\item nb = number of burns: default nb = 300000
\item nc = number of chains: default nc = 3
}}
}
\value{
fitted model. A list of class "JagsUI".
}
\description{
\code{\link[=tor_fit]{tor_fit()}} fits a binomial mixture model using Bayesian
inference. It uses Rjags in the background and enables users to specify
some - but not all - sampling parameters. The structure of the model can also
be changed. Users who want more flexibility are encouraged to use Rjags
directly. The function considers the assumed relation between metabolic rate (MR) and ambient temperature (Ta)
(Speakman & Thomas 2003). In the hypothermic state (torpor) and above some
threshold Ta (Tmin), MR follows an exponential curve reflecting the Arrhenius
rate enhancing effect of temperature on chemical reactions, whereas below
Tmin, it increases linearly with decreasing Ta to maintain a minimal Tb in
torpor. In the euthermic state, MR solely increases linearly with decreasing Ta.
More information about the model can be found in the vignette \code{vignette("model description", package = "torpor")}.
}
\details{
This model should be applied with sufficient sample size and if
evidence suggest that individuals under study will conform to the previously
described pattern while in torpor.
}
\examples{
rm(list = ls())
data(test_data3)
test2 <- tor_fit(MR = test_data3[,2],
Ta = test_data3[, 1],
BMR = 1.005,
TLC = 29,
model = NULL,
fitting_options = list(ni = 5000, nb = 3000, nc = 2))
}
\seealso{
Other fit: \code{\link{complete_args}}
}
\concept{fit}
