% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_summarise.R
\name{tor_overlap}
\alias{tor_overlap}
\title{Check priors/posteriors overlap}
\usage{
tor_overlap(tor_obj)
}
\arguments{
\item{mod}{a fitted model from \code{\link[=estimate_parameters]{estimate_parameters()}}}
}
\value{
a data.frame
}
\description{
\code{\link[=tor_overlap]{tor_overlap()}} generates prior/posterior overlap values for Tlc,
Tt and Betat. Values larger than 0.3 should lead to the conclusion that conforming
torpor, regulated torpor or thermoregulation respectively could not be modeled
with the data provided (Fasel et al. (in prep)). \code{\link[=tor_overlap]{tor_overlap()}} can be used independently but is also used internally in
\code{\link[=tor_summarise]{tor_summarise()}}.
}
\seealso{
Other summary: 
\code{\link{get_parameters}()},
\code{\link{tor_summarise}()}
}
\concept{summary}
