% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_plot.R
\name{tor_plot}
\alias{tor_plot}
\title{Plot raw data and predicted values of a fitted model}
\usage{
tor_plot(
  tor_obj = NULL,
  plot_type = "base",
  legend = TRUE,
  col_torp = "cornflowerblue",
  col_euth = "coral3",
  col_Mtnz = "black",
  ylab = "M",
  xlab = "Ta",
  pdf = FALSE
)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}

\item{plot_type}{A character string specifying the type of plot desired. Either "base" or "ggplot". Note that ggplots are still in development stage.}

\item{legend}{A logical specifying if a legend should be added to the plot. Only work for base plot at the moment.}

\item{col_torp}{color for torpor model prediction and points}

\item{col_euth}{color for euthermy model prediction and points}

\item{col_Mtnz}{color of Mtnz}

\item{ylab}{y label}

\item{xlab}{x label}

\item{pdf}{logical if a .pdf copy of the plot should be saved}
}
\value{
a base-R plot or a ggplot object
}
\description{
\code{\link[=tor_plot]{tor_plot()}} provides a plot of the metabolic rate (M) values over the respective
ambient temperature (Ta).
Raw data and predicted values are presented in different colors depending
of the metabolic state. Predicted values are represented by continuous and
stripped lines for the estimates’ median and 95\% CI bounds of the posterior
distribution, respectively.
For more flexibility the users can use \code{\link[=tor_fit]{tor_fit()}} and \code{\link[=tor_predict]{tor_predict()}} directly.
}
