% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_predict}
\alias{tor_predict}
\title{get Prediction}
\usage{
tor_predict(mod, Ta)
}
\arguments{
\item{mod}{a fitted model from the function \code{\link[=tor_fit]{tor_fit()}}}

\item{Ta}{a vector of temperatur for which the prediction should be made}
}
\value{
a data frame with predicted values
}
\description{
The function \code{\link[=tor_predict]{tor_predict()}} provides the predicted MR and 95CI bounds at
a given Ta, in euthermic and/or torpid stage.
}
\examples{
data(test_data2)
test_mod <- tor_fit(MR = test_data2[,2],
Ta = test_data2[, 1],
BMR = 1.49,
TLC = 28.88,
fitting_options = list(nc = 2, nb = 3000, ni = 5000))
tor_predict(mod = test_mod, Ta = 1:20)
}
\seealso{
Other predict: \code{\link{eut_predict_fun}},
  \code{\link{tor_classify}}, \code{\link{tor_predict_fun}}
}
\concept{predict}
