% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_predict.R
\name{tor_predict}
\alias{tor_predict}
\title{Model predictions}
\usage{
tor_predict(tor_obj, Ta)
}
\arguments{
\item{tor_obj}{a fitted model from \code{\link[=tor_fit]{tor_fit()}}}

\item{Ta}{a vector of temperature}
}
\value{
a data.frame
}
\description{
\code{\link[=tor_predict]{tor_predict()}} provides the predicted metabolic rate (M) and 95\% CI bounds at
a given ambient temperature (Ta), in euthermic and/or torpid state.
}
\examples{
\dontrun{
test_mod <- tor_fit(M = test_data2[,2],
                   Ta = test_data2[, 1],
                   fitting_options = list(nc = 1, nb = 3000, ni = 5000))
tor_predict(tor_obj, Ta = 10:35)
}
}
\seealso{
Other predict: 
\code{\link{eut_predict_fun}()},
\code{\link{tor_classify}()},
\code{\link{tor_predict_fun}()}
}
\concept{predict}
